Add Sidekiq job for availability message processing
